Add a warning to fork the repository
This page is still a work in progress. In the future there might be new
tests added. It would be pretty bad, if we change the wording on the test
and you deliver a solution that is not up to date. Because of that, it's
a good idea to copy the text before starting to work.